// frontend/src/app/modules/landing/employer/models/profile-edit.ts
export interface ProfileEdit {
    organization_name: string;
    organization_details: string;
    logo: any;
    industry: any;
    location: string;
    address_1: string;
    address_2: string;
    organization_website_url: string;
    organization_upper_limit: string;
    organization_lower_limit: string;
}